+++ packages/core/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 This document will contain the list of changes made to the framework. The format is based on the [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/) specification.
 
+## Version 1.3.6
+* feat(exceptions): add new exception code `E_GEN_INVALID_ARGUMENT_TYPE`
+* feat(exceptions): add public enum with all HTTP codes exported as `httpCode` and typedef `TiHttpCode`. Use it as the type for the `httpCode` exception property
+* feat(tools): add `description` property to enum objects
+* feat(tools): add `contains` property to enum objects
+* feat(tools)!: change enum factory behavior to create a copy of the seed object instead of modifying it (may break code relying on seed object mutation)
+* refactor(exceptions): invalid `httpCode` inputs are coerced to `undefined`; `Exception.asJSON` includes `httpCode` only when defined
+* refactor(tools): deprecate method `getEnumName` in favor of `enum.name`; removal planned for 1.4.0
+* refactor(tools)!: remove method `createCSVFile` as it is unnecessary for the framework's operation. It also eliminates the dependency from `fs-extra` package
+* refactor(tools)!: disallow reserved keys as enum names; make enum properties non-enumerable and freeze the enum object to enforce immutability (may affect code relying on enumeration/extension)
+* fix(tools): fix a bug in the `RetryPolicy` class that caused the `maxAttempts` property to be ignored
+* build(npm): update npm dependencies to their latest versions
+* build(npm): remove `fs-extra` package as it is not used by the framework
+* build(npm)!: bump the minimum supported Node.js version to 18.0.0
+
 ## Version 1.3.5
 * fix(redis integration): fix a duplicated log entry on connection ready event if multiple observers are registered
 * feat(service caller)!: change log level of error result in `process` method from `ERROR` to `DEBUG`. Implementers are expected to handle this and decide if the error should be propagated further or not